feat: replace on ffmpeg_kit

+++ lib/extensions/statistics_extension.dart
@@ -1,9 +1,9 @@
-import 'package:ffmpeg_helper/ffmpeg_helper.dart';
+import 'package:ffmpeg_kit_flutter/statistics.dart';
 
 extension OnStatistics on Statistics {
   /// get time by duration
   Duration getTimeDuration() {
-    return Duration(microseconds: getTime());
+    return Duration(microseconds: getTime().toInt());
   }
 
   /// get time by milliseconds
